lib/Audio: update async S/PDIF input + Resampler to alex6679 current — optical-in -> TAC5212 DAC works
Replaces our stale vendored snapshot of alex6679's teensy-4-spdifIn
(async_input_spdif3, Resampler, Quantizer, biquad) with his CURRENT code. The
old version hung the firmware at init whenever the S/PDIF receiver came up
alongside the SAI (codec) output — the actual root cause of the optical-input
"dark / reboot loop" failures. It was never the board or the silicon; we were a
library version behind. His current async input (configure() -> newConfiguration,
Resampler::updateIncrement/setPos, ARM_DWT_CYCCNT timing) coexists with the SAI.

Proven:
- projects/spike_spdif_alex: his code + AudioOutputI2S -> runs, no hang.
- projects/spike_spdif_alex_dac: full optical loop CONFIRMED BY EAR — 1 kHz tone
  -> S/PDIF OUT -> loopback -> S/PDIF IN -> AudioOutputI2S -> TAC5212 DAC (I2S,
  HpDriver, DIN/DOUT-swap bodge) -> amp.

Also in this commit:
- lib/OpenAudio .../input_spdif3_f32.cpp: define AudioInputSPDIF3_F32::sample_rate_Hz
  (declared-but-undefined -> linker error; the sync F32 input had never built).
- projects/spike_spdif_min: minimal stock-Audio S/PDIF out + loopback bring-up test.
- projects/t-dsp_spdif_monitor: F32 codec monitor + full bring-up log/recipe (NOTES.md)
  incl. the DAC operational recipe (tdmOut-first update_responsibility, SPDIF-out
  clock for PLL lock, DOUT-disable bodge, HpDriver).
- lib/TAC5212/BOARD_NOTES.md: DIN/DOUT swap+short bodge, TCA9544A mux 0x70 ch3, SHDNZ.

CAVEAT: the new Resampler API breaks lib/OpenAudio's F32 async wrapper
(async_input_spdif3_F32.cpp still calls the old API: addToSampleDiff/addToPos/
fixStep). The spikes lib_ignore OpenAudio. t-dsp_f32_audio_shield optical-in needs
the F32 wrapper ported to alex's current logic — see
projects/spike_spdif_alex_dac/NOTES.md.

+++ lib/Audio/Quantizer.h
@@ -29,7 +29,7 @@
 #ifndef quantizer_h_
 #define quantizer_h_
 
-#include <Arduino.h>     // github.com/PaulStoffregen/cores/blob/master/teensy4/Arduino.h
+#include "Arduino.h"
 
 //#define DEBUG_QUANTIZER
 
@@ -40,9 +40,9 @@ public:
     ///@param audio_sample_rate currently only 44.1kHz and 48kHz are supported
     Quantizer(float audio_sample_rate);
     void configure(bool noiseShaping, bool dither, float factor);
-    void quantize(float* input, int16_t* output, uint16_t length);
+    void quantize(float* input, int16_t* output, int32_t length);
     //attention outputInterleaved must have length 2*length
-    void quantize(float* input0, float* input1, int32_t* outputInterleaved, uint16_t length);
+    void quantize(float* input0, float* input1, int32_t* outputInterleaved, int32_t length);
     void reset();
         
 private:
